requirements.txt
astroid==1.4.5
colorama==0.3.7
cycler==0.10.0
eventlet==0.19.0
Flask==0.10.1
greenlet==0.4.9
itsdangerous==0.24
javabridge==1.0.14
Jinja2==2.8
lazy-object-proxy==1.2.2
liac-arff==2.1.0
MarkupSafe==0.23
matplotlib==1.5.1
numpy==1.11.0
pandas==0.18.1
PIL==1.1.7
pyflakes==1.2.3
pylint==1.5.5
pyparsing==2.1.4
python-dateutil==2.5.3
python-weka-wrapper==0.3.8
pytz==2016.4
scikit-learn==0.17.1
scipy==0.17.1
six==1.10.0
unbalanced-dataset==0.1.dev0
Werkzeug==0.11.9
wrapt==1.10.8
